requirements.txt
python-dotenv>=1.0.0
openai>=1.40.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
pandas>=2.2.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
pyyaml>=6.0.2
